requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests==2.31.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
langchain==0.0.350
langchain-openai==0.0.2
langchain-community==0.0.10
httpx==0.25.2
python-dotenv==1.0.0
websockets==12.0
python-dateutil==2.8.2
google-api-python-client==2.109.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-auth==2.23.4
google-cloud-core==2.3.3
sqlalchemy==1.4.48
alembic==1.12.1
psycopg2-binary==2.9.7
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.1.0
